websocket/client.go: add our own close handler that intercepts ws.ErrCloseSent.
In Gorilla WebSockets version 1.5.1, they introduced a change where the error
code from the `WriteControl` function call is returned in their default close handler.

This is generally acceptable; however, when the connection initiates a close
handshake, all writes to the connection result in an `ErrCloseSent` error.

This means that instead of receiving a `CloseError` with the actual code and
message from the other end, their close handler returns `ErrCloseSent`.